moved new fs-related function into it's own module; extends fs with the new functions move, fileExists and pathExists; also fixed SaveFileList again (missing this);
